+++ gulp/tasks/styles.js
@@ -1,5 +1,6 @@
 const { mkdir, writeFile } = require('fs/promises')
-const { dirname, join, parse } = require('path')
+const { dirname, join, parse, relative } = require('path')
+const { fileURLToPath } = require('url')
 
 const autoprefixer = require('autoprefixer')
 const cssnano = require('cssnano')
@@ -20,8 +21,8 @@ function compileStyles(assetPath, { srcPath, destPath, output = {} }) {
   const to = join(destPath, output.file ?? `${name}.css`)
 
   const taskFn = async () => {
-    const { css } = await compileAsync(from, {
-      loadPaths: ['node_modules'],
+    const { css, sourceMap } = await compileAsync(from, {
+      loadPaths: ['.', 'node_modules'],
       quietDeps: true,
       silenceDeprecations: [
         'color-functions',
@@ -29,14 +30,38 @@ function compileStyles(assetPath, { srcPath, destPath, output = {} }) {
         'import',
         'mixed-decls',
         'slash-div'
-      ]
+      ],
+      sourceMap: true,
+      sourceMapIncludeSources: true
     })
 
+    // Make source file:// paths relative
+    if (sourceMap?.sources) {
+      sourceMap.sources = sourceMap.sources.map((path) =>
+        path.startsWith('file:') ? relative(from, fileURLToPath(path)) : path
+      )
+    }
+
+    // Apply PostCSS transforms (e.g. vendor prefixes)
     const processor = postcss([autoprefixer(), cssnano()])
-    const result = await processor.process(css, { from, to })
+    const result = await processor.process(css, {
+      from,
+      to,
+
+      // Include Sass sources
+      map: {
+        annotation: true,
+        inline: false,
+        prev: sourceMap
+      }
+    })
 
     await mkdir(dirname(to), { recursive: true })
     await writeFile(to, result.css)
+
+    if (result.map) {
+      await writeFile(`${to}.map`, result.map.toString())
+    }
   }
 
   taskFn.displayName = 'compile:styles'